refactor: 충돌 방지를 위한 일부 수정
